requirements.txt
pyvmomi
graphitesend
pyyaml
